finish refactoring move generator (restores build and passes all tests)

+++ dctl/evaluate/delta.hpp
@@ -50,7 +50,7 @@ private:
         template<typename Position>
         static int mobility(Position const& p)
         {
-                return Active::mobility(p) - Passive::mobility(p);
+                return static_cast<int>(Active::mobility(p) - Passive::mobility(p));
         }
 };
